Add serialization support (#1)
This PR creates the new trait `Distance` which has one method with the
following signature:
```rust
pub trait Distance<T: ?Sized> {
    fn distance(&self, a: &T, b: &T) -> isize;
}
```
The functions `hamming_distance()` and `levenshtein_distance()` are
refactored to `impl` the Distance trait for the new structs
`HammingDistance` and `LevenshteinDistance`. These structs do not hold
state, but can be passed in as type parameters to create `BkTree`
objects with known size at compile time, e.g.
```rust
let mut bk = BkTree::new(HammingDistance);
// ...
let mut bk = BkTree::new(LevenshteinDistance);
```

The remainder of the refactor changes consist in refactoring `BkTree` to
accept the second type parameter and fixing the unit tests accordingly.

Finally, this PR adds the `serde-support` feature to enable serializing
and deserializing `BkTree` objects for those that would desire to do so.

This constitutes a breaking change to the API of the `BkTree`. Adhering
to [semver guidelines](https://semver.org), the major version of this
crate has been incremented.

Note that this PR leans heavily on the work done by @eugene-bulkin in
the [rust-bk-tree](https://github.com/eugene-bulkin/rust-bk-tree) crate
as well as the work on an [outstanding PR to support
serialization](https://github.com/eugene-bulkin/rust-bk-tree/pull/12) in
that crate by @stnbu

---------

Co-authored-by: IGI-111 <igi-111@protonmail.com>